Move to a dedicated `char_traits` header
